% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{A Self-made Function}
\usage{
clean_data(data)
}
\arguments{
\item{data}{A dataframe}
}
\description{
This function allows you to remove punctuation marks, loanwords and numbers from your data.
}
\examples{
clean_data()
}
\keyword{clean}
\keyword{data}
