% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_grammar_level.R
\name{match_grammar_level}
\alias{match_grammar_level}
\title{A Self-made Function}
\usage{
match_grammar_level(data)
}
\arguments{
\item{data}{A dataframe}
}
\description{
This function allows you to match the grammar level of your data.
}
\examples{
match_grammar_level()
}
\keyword{grammar}
\keyword{level}
\keyword{match}
