% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_vocabulary_level.R
\name{match_vocabulary_level}
\alias{match_vocabulary_level}
\title{A Self-made Function}
\usage{
match_vocabulary_level(data)
}
\arguments{
\item{data}{A dataframe}
}
\description{
This function allows you to match the vocabulary level of your data.
}
\examples{
match_vocabulary_level()
}
\keyword{level}
\keyword{match}
\keyword{vocabulary}
